v1.0.1.5 Get anyone's tournament w/o subdomain
Added a field for tournament name to Challonge Loader. Providing the
tournament name/id will allow user to import a challonge bracket that is
neither associated with the provided api key nor part of any subdomain.

ChallongeImporter's import method now takes an extra argument for
tournament name. If provided, that tournament will be looked up and added
to the list of possible imports.

In the ChallongeApiWrapper, the ShowTournament method now takes a string
instead of an int. The int gets cast into a string anyway, and the
tournament name or id can be used. ShowTournament also now handles the
case where a subdomain and tournament name are both provided. The
Tournament class has a slight modification, it now has a specified equals
behavior (using full url), so that duplicates aren't added to the list if
name of a tournament is specified along with its subdomain or associated
api key.