Correction of Toolbox.cs for Decay Problems
This edit of Toolbox.cs attempts to correct the problems with the decay interval.  Namely the fact that a 7 day decay and a 1 week decay do not yield the same scores as well as unexpected decay do to how decayDays and decayMonths are retained.  There are more thorough comments in the parts that I have changed.