Change offending smash.gg API properties from bool to object
According to Smashboards user Demandatwin, these are the offending properties. ThreeFold changed bool to object, while Demandatwin changed the bool to string. Might just need to change all bool properties if we're not using the JSON.NET package to deserialize. But since the smash.gg API will probably change again and break things, this is just a stopgap.